Several testsuite fixes/enhancments
* test/getmntent.c: I removed the checks which called endmntent()
  multiple times resp. with a NULL pointer.  Behavior in first
  case is undefined and latter causes a segfault (I could not
  find documentation that this is allowed).  Hence, it does not
  make sense to test this.

* test/stdlib/tst-malloc.c: test 'malloc(0)' only, when
  WANT_MALLOC_ZERO is set

* test/time/tst-strftime.c: the '%Zu' printf format specifier is
  bogus (from libc5 ages). I replaced it with the '%zu' one.

  Please note, that this test still fails because it uses glibc
  specific features.

* test/time/tst-strptime.c: removed tests which set a non-C
  locale. Broken usage of 'strptime()' result was fixed too.
  This test should succeed now.

* test/math.c: this test was added while working on the '%F/%G'
  patch and verifies the isinf() + isnan() functions.

* fixes lot of broken printf() format specifiers

* fixes tests using isinf(); the math.c test checks the dietlibc
  implementation of isinf() (which differs between positive and
  negative inf). All other tests use boolean retval only and test
  for the signess of the values

+++ test/mktime.c
@@ -9,15 +9,15 @@ int main() {
   t.tm_mday=29;
   t.tm_mon=2;
   t.tm_year=100;
-  printf("%d\n",mktime(&t));
+  printf("%ld\n",(long)mktime(&t));
   t.tm_mday=1;
   t.tm_mon=3;
   t.tm_year=102;
-  printf("%d\n",mktime(&t));
+  printf("%ld\n",(long)mktime(&t));
   t.tm_mday=1;
   t.tm_mon=6;
   t.tm_year=102;
-  printf("%d\n",mktime(&t));
+  printf("%ld\n",(long)mktime(&t));
   return 0;
 }
